LayoutTests/media/media-source/media-source-append-buffer-full-quota-exceeded-error-onstart-expected.txt


EVENT(sourceopen)
Appending buffer
EXPECTED (exception == 'QuotaExceededError: The quota has been exceeded.') OK
END OF TEST
